Store and fetch the bakery outside of the Redux store.
