+++ src/@canonical/macaroon-bakery.d.ts
@@ -1,0 +1,1 @@
+declare module "@canonical/macaroon-bakery";
