+++ src/app/action-types.ts
@@ -1,0 +1,12 @@
+// Action labels - these are defined in a separate file to prevent circular imports.
+export const actionsList = {
+  logOut: "LOG_OUT",
+  connectAndPollControllers: "CONNECT_AND_POLL_CONTROLLERS",
+  storeConfig: "STORE_CONFIG",
+  storeLoginError: "STORE_LOGIN_ERROR",
+  storeUserPass: "STORE_USER_PASS",
+  storeVersion: "STORE_VERSION",
+  storeVisitURL: "STORE_VISIT_URL",
+  updateControllerConnection: "UPDATE_CONTROLLER_CONNECTION",
+  updatePingerIntervalId: "UPDATE_PINGER_INTERVAL_ID",
+};
