+++ src/app/root.js
@@ -1,7 +1,9 @@
 import immerProduce from "immer";
 import cloneDeep from "clone-deep";
 
-import { actionsList } from "./actions";
+import bakery from "app/bakery";
+
+import { actionsList } from "./action-types";
 
 function rootReducer(state = {}, action) {
   return immerProduce(state, (draftState) => {
@@ -10,9 +12,6 @@ function rootReducer(state = {}, action) {
         const connections = cloneDeep(state.controllerConnections || {});
         connections[action.payload.wsControllerURL] = action.payload.info;
         draftState.controllerConnections = connections;
-        break;
-      case actionsList.storeBakery:
-        draftState.bakery = action.payload;
         break;
       case actionsList.storeConfig:
         draftState.config = action.payload;
@@ -32,7 +31,7 @@ function rootReducer(state = {}, action) {
         draftState.visitURL = action.payload;
         break;
       case actionsList.logOut:
-        delete draftState.bakery.storage._store.identity;
+        delete bakery.storage._store.identity;
         delete draftState.controllerConnections;
         delete draftState.visitURL;
         break;
