+++ src/juju/action-types.ts
@@ -1,0 +1,11 @@
+// Action labels - these are defined in a separate file to prevent circular imports.
+export const actionsList = {
+  clearControllerData: "CLEAR_CONTROLLER_DATA",
+  clearModelData: "CLEAR_MODEL_DATA",
+  updateControllerList: "UPDATE_CONTROLLER_LIST",
+  updateModelInfo: "UPDATE_MODEL_INFO",
+  updateModelStatus: "UPDATE_MODEL_STATUS",
+  updateModelList: "UPDATE_MODEL_LIST",
+  processAllWatcherDeltas: "PROCESS_ALL_WATCHER_DELTAS",
+  populateMissingAllWatcherData: "POPULATE_MISSING_ALLWATCHER_DATA",
+};
