+++ src/juju/actions.js
@@ -2,17 +2,7 @@ import cloneDeep from "clone-deep";
 
 import { fetchAndStoreModelStatus } from "juju";
 
-// Action labels
-export const actionsList = {
-  clearControllerData: "CLEAR_CONTROLLER_DATA",
-  clearModelData: "CLEAR_MODEL_DATA",
-  updateControllerList: "UPDATE_CONTROLLER_LIST",
-  updateModelInfo: "UPDATE_MODEL_INFO",
-  updateModelStatus: "UPDATE_MODEL_STATUS",
-  updateModelList: "UPDATE_MODEL_LIST",
-  processAllWatcherDeltas: "PROCESS_ALL_WATCHER_DELTAS",
-  populateMissingAllWatcherData: "POPULATE_MISSING_ALLWATCHER_DATA",
-};
+import { actionsList } from "./action-types";
 
 // Action creators
